Standardize old resources to to use list method with SQL Injection protection and stardardize PollingProfile file (#2839)
* wip: update list on demo and environment to use the same standards as tests and transactions

* Fixing provisioning test

* Update provisioning test
